+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v3.12.4
+### Bug Fixes
+* Strip material swap bindings that point to slots that don't exist on the renderer.
+  * Merging meshes could result in those slots now existing and getting overwritten by the material swap animation.
+
 ## v3.12.3
 ### Changes
 * Add ignore warning 3554 to generated shaders (The attribute is unknown or invalid for the specified statement).